!B (CE-13801) (Wavicle) Fixed EnableByConfig feature. !B (Wavicle) Changed ChaosKeyV to match recent ChaosKey change, and fixed unit test. Fixed assert in pfx1 TextureTile.Correct. !F (Wavicle) Added IParticleEffect.CreateAttributeInstance.
Copied from Perforce
 Change: 1592528

+++ Code/CryEngine/CryCommon/CryParticleSystem/ParticleParams.h
@@ -738,9 +738,9 @@ struct ParticleParams
 
 		void Correct()
 		{
-			nFirstTile = std::min<uint8>(nFirstTile, nTilesX * nTilesY - 1);
-			nAnimFramesCount = std::min<uint8>(nAnimFramesCount, GetTileCount());
-			nVariantCount = std::min<uint8>(nVariantCount, GetTileCount() / nAnimFramesCount);
+			nFirstTile = std::min<uint>(nFirstTile, nTilesX * nTilesY - 1);
+			nAnimFramesCount = std::min<uint>(nAnimFramesCount, GetTileCount());
+			nVariantCount = std::min<uint>(nVariantCount, GetTileCount() / nAnimFramesCount);
 		}
 
 		AUTO_STRUCT_INFO;
